refactor: remove remaining command dependency bags
